backend/requirements.txt
fastapi
uvicorn
python-docx
PyMuPDF
python-dotenv
sqlalchemy
python-multipart
httpx
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
reportlab
